pyproject.toml
[tool.poetry]
name = "asmre-backend"
version = "0.0.1"
description = ""
authors = ["XYCode Kerman <xycode-xyc@outlook.com>"]
license = "LGPL-v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
odmantic = "^1.0.2"
dnspython = "^2.6.1"
pyjwt = "^2.8.0"
pytest-asyncio = "^0.23.6"
asyncer = "^0.0.7"
requests = "^2.31.0"
casbin-rabbitmq-watcher = "^1.3.0"
casbin-sqlalchemy-adapter = "^1.1.0"
pymysql = "^1.1.0"
cryptography = "^42.0.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"
mypy = "^1.10.0"
types-pymysql = "^1.1.0.20240425"


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
